requirements.txt
certifi==2018.10.15
asgiref==3.2.7
Django==3.0.5
django-crispy-forms==1.9.0
django-filter==2.2.0
django-sslserver==0.22
django-widget-tweaks==1.4.5
PyMySQL==0.9.3
mysqlclient==1.4.6
SQLAlchemy==1.3.15
gunicorn==20.0.4
pip==20.0.2
uuid==1.30
setuptools==39.0.1
Pillow==7.1.2
html5lib==1.0.1
pyPdf2==1.26.0
reportlab==3.5.42
webencodings==0.5.1
xhtml2pdf==0.2.4
wheel==0.32.2
